+++ system/src/main/java/io/github/sds100/keymapper/system/apps/AndroidPackageManagerAdapter.kt
@@ -2,6 +2,7 @@ package io.github.sds100.keymapper.system.apps
 
 import android.annotation.SuppressLint
 import android.app.ActivityOptions
+import android.app.KeyguardManager
 import android.app.PendingIntent
 import android.content.ActivityNotFoundException
 import android.content.BroadcastReceiver
@@ -48,6 +49,10 @@ class AndroidPackageManagerAdapter @Inject constructor(
     private val ctx: Context = context.applicationContext
     private val packageManager: PackageManager = ctx.packageManager
 
+    private val keyguardManager: KeyguardManager by lazy {
+        ctx.getSystemService(KeyguardManager::class.java)
+    }
+
     private val fetchPackages: MutableSharedFlow<Unit> = MutableSharedFlow()
     override val onPackagesChanged: MutableSharedFlow<Unit> = MutableSharedFlow()
     override val installedPackages = MutableStateFlow<State<List<PackageInfo>>>(State.Loading)
@@ -62,7 +67,7 @@ class AndroidPackageManagerAdapter @Inject constructor(
                 Intent.ACTION_PACKAGE_ADDED,
                 Intent.ACTION_PACKAGE_REMOVED,
                 Intent.ACTION_PACKAGE_REPLACED,
-                -> {
+                    -> {
                     coroutineScope.launch {
                         fetchPackages.emit(Unit)
                         onPackagesChanged.emit(Unit)
@@ -233,10 +238,10 @@ class AndroidPackageManagerAdapter @Inject constructor(
         val leanbackIntent = packageManager.getLeanbackLaunchIntentForPackage(packageName)
         val normalIntent = packageManager.getLaunchIntentForPackage(packageName)
 
-        val intent = leanbackIntent ?: normalIntent
+        val packageIntent = leanbackIntent ?: normalIntent
 
         // intent = null if the app doesn't exist
-        if (intent == null) {
+        if (packageIntent == null) {
             try {
                 val appInfo = ctx.packageManager.getApplicationInfo(packageName, 0)
 
@@ -250,11 +255,17 @@ class AndroidPackageManagerAdapter @Inject constructor(
                 return KMError.AppNotFound(packageName)
             }
         } else {
-            val pendingIntent = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-                PendingIntent.getActivity(ctx, 0, intent, PendingIntent.FLAG_IMMUTABLE)
+            // Use a trampoline activity that will dismiss the keyguard when it is locked.
+            val intent = if (keyguardManager.isKeyguardLocked) {
+                Intent(ctx, TrampolineActivity::class.java).apply {
+                    putExtra(TrampolineActivity.EXTRA_INTENT, packageIntent)
+                }
             } else {
-                PendingIntent.getActivity(ctx, 0, intent, 0)
+                packageIntent
             }
+
+            val pendingIntent =
+                PendingIntent.getActivity(ctx, 0, intent, PendingIntent.FLAG_IMMUTABLE)
 
             if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.UPSIDE_DOWN_CAKE) {
                 val bundle = ActivityOptions.makeBasic()